+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 - Added account endpoint
 - Added account program endpoint
 - Added account region endpoint
+- Added account user endpoint